requirements.txt
chromadb
langchain-core
langchain-ollama
langchain-community
pandas